If no connection when new high score, repeated attempts to send are made
Locally a variable indicating the synchronization status of local and server information is stored (whether the local highest score has been successfully sent to the server).
Sending will start a coroutine that will attempt to connect with the server after a set delay. The status of the synchronization variable is checked when starting the game;
if the app was exited when the high score has not been sent to  the server, more attemps are made after starting.