Fixed freeze when winning + UI change
Turku sprite in use does not have the guide text. Instead the text is an element in the UI canvas. Zoomed a bit farther to leave room between the time bar and the northmost corner of Turku.
Fix to the freeze was setting a single object reference in the scene (particle system).