Finland is positioned in the center bottom + success is conveyed without words
Instead of a message, the color of the city circle that is revealed at the indicates whether the player succeeced. (Green for success, red for failure).
The UI status tests were removed.

Also fixed a bug: the correct city was not revealed if the timer ran out. Now it is.