Dev cheats for winning/losing work with the existing games
If there is a MonoBehavior script in a minigame scene that implements IMinigameEnder interface, the cheat script finds a reference to this at the beginning of the minigame.
The cheat script will call the methods WinMinigame or LoseMinigame (required by the interface) in response to 'W' or 'L' keys being pressed.

The buttons could be closer together, and e.g. in Turku game there are unnecessarily many methods to quit the minigame